Add files via upload
This book is partially based on previous work, which I will briefly mention here and in
a more detailed bibliography at the end of this book. In particular, there was a two-part
article on the WinHelp .HLP file format I wrote for Dr. Dobb's Journal, an article in
Windows Developer's Journal (WDJ) on the .SHG and .MRB file formats, and another
article in WDJ on the file format and LZ77 algorithm used by COMPRESS.EXE. The
.ANN file format appeared in PC Magazine (Volume 14, No. 15) in an article I
co-wrote with Jim Mischel on advanced WinHelp techniques.
Two chapters in this book came from work done by other people. The . RES file format was originally uncovered by Dmitry M. Rogatkin and Alex G. Fedorov and published in Dr. Dobb's Journal in August 1993. The .PI F file format was
reverse-engineered by Michael P. Maurice and was also published in Dr. Dobb's Journal in July 1993.
The W4 file format presented in Chapter 9 was provided by Clive Turvey. He and
several other people figured out the format, and Clive, as far as I know, was the one
that figured out the compression algorithm. He was the first I heard of to write his own
routines for decompressing W4 files, whereas other people were calling directly into
the Double Space decompression routines.
The rest of the files covered in this book were solved by Mike and I and have not
been previously published.
Information on the .HLP file format in this book has been completely revised
from the original articles. A lot of corrections and pieces of missing information
have been filled in.